API Completed, write unit tests

+++ ConsoleApp2/AssembleError.cs
@@ -9,8 +9,8 @@ namespace AssemblerSimulator8085
     public class AssembleError
     {
         public string error { get; set; }
-        public ushort line_no { get; set; }
+        public int line_no { get; set; }
 
-        public AssembleError(string error_name, ushort line_number) => (error, line_no) = (error_name, line_number);
+        public AssembleError(string error_name, int line_number) => (error, line_no) = (error_name, line_number);
     }
 }
